refactor renderer by page Fixes #58
