+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
-dist
-out/
+dist/*
+out/*
 test
+html/js/*
