+++ src/render/pages/_shared/inputs.ts
@@ -1,0 +1,8 @@
+export function getInputValue(id: string): string | undefined {
+  const i = document.querySelector(`#contents #${id}`) as HTMLInputElement;
+  return i?.value;
+}
+export function setInputValue(id: string, value: string):void {
+  const i = document.querySelector(`#contents #${id}`) as HTMLInputElement;
+  i.value = value;
+}
